+++ src/main.rs
@@ -726,6 +726,15 @@ impl App {
             "General Overview {#general_overview}",
             &stored_static.html_report_flag,
         );
+        match stored_static.config.action {
+            Some(Action::CsvTimeline(_)) | Some(Action::JsonTimeline(_)) => {
+                println!();
+                println!("Please report any issues with Hayabusa rules to: https://github.com/Yamato-Security/hayabusa-rules/issues");
+                println!("Please report any false positives with Sigma rules to: https://github.com/SigmaHQ/sigma/issues");
+                println!("Please submit new Sigma rules with pull requests to: https://github.com/SigmaHQ/sigma/pulls");
+            }
+            _ => {}
+        }
 
         // Qオプションを付けた場合もしくはパースのエラーがない場合はerrorのstackが0となるのでエラーログファイル自体が生成されない。
         if ERROR_LOG_STACK.lock().unwrap().len() > 0 {